♻️ Update SMS status mappings for Twilio
Until now, a successfully sent SMS message would appear in the admin UI
as "Sent to an international number", even when the recipient is in the
UK. Furthermore, if/when Twilio reported a message as "undelivered", the
SMS status would fail to update in our database since we didn't have a
mapping for this particular status.

This commit updates our mapping of SMS callbacks, removing the mention
of international numbers and adding a mapping for an "undeliverable"
callback.