🐛 Fix linting failures in updated healthcheck.py
We recently made some changes to how the health of the application is
tested, removing dependencies on the database and focusing purely on
whether the Flask app is reachable. As part of these changes we
introduces a handful of linting failures which are detected by `flake8`
and currently cause our CI runs to fail.

This commit removes some unusued imports and adds extra line-spacing
between the imports and method definition.